Separate Servers and Processes features namespaces
