+++ ArmaForces.Arma.Server.Tests/Helpers/TestDedicatedServer.cs
@@ -2,8 +2,8 @@
 using System.Threading;
 using System.Threading.Tasks;
 using ArmaForces.Arma.Server.Features.Modsets;
-using ArmaForces.Arma.Server.Features.Server;
-using ArmaForces.Arma.Server.Features.Server.DTOs;
+using ArmaForces.Arma.Server.Features.Servers;
+using ArmaForces.Arma.Server.Features.Servers.DTOs;
 using AutoFixture;
 using CSharpFunctionalExtensions;
