+++ ArmaForces.Arma.Server/Features/Processes/IArmaProcess.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using ArmaForces.Arma.Server.Features.Parameters;
 using CSharpFunctionalExtensions;
 
-namespace ArmaForces.Arma.Server.Features.Server
+namespace ArmaForces.Arma.Server.Features.Processes
 {
     public interface IArmaProcess
     {
