+++ ArmaForces.Arma.Server/Features/Processes/IArmaProcessDiscoverer.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace ArmaForces.Arma.Server.Features.Server
+namespace ArmaForces.Arma.Server.Features.Processes
 {
     public interface IArmaProcessDiscoverer
     {
