+++ ArmaForces.ArmaServerManager.Tests/Providers/Server/ServerProviderTests.cs
@@ -1,7 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using ArmaForces.Arma.Server.Features.Modsets;
-using ArmaForces.Arma.Server.Features.Server;
+using ArmaForces.Arma.Server.Features.Processes;
+using ArmaForces.Arma.Server.Features.Servers;
 using ArmaForces.Arma.Server.Tests.Helpers;
 using ArmaForces.ArmaServerManager.Providers;
 using ArmaForces.ArmaServerManager.Providers.Server;
